Adds safe save method with optimistic locking
Provides a convenient way to prevent concurrent and overwriting changes
when saving MalwareScanFileState objects.
A new save method is exposed in PersistenceRepository interface that
receives the expected current value to be overwritten with a new value.
If that current value was changed by another thread, the save method
will throw a PersistenceException with an OptimisticLockException as
the cause.

This change also introduces a version attribute for the
MalwareScanFileState model object.

Ref: DLP-5043, DLP-5044