Changes the MS OperationPersistenceService to use putIfAbsent
The create method will use 'putIfAbsent' instead of simply 'put'. This will
prevent to overwrite any existing State by mistake.
The method will succeed and return a response with SUCCESS status only if the
given MalwareScanFileState in the request was saved. If it was not possible
to save the new state due a previously existing record, then the method
will return STATE_PRESENT.

With this change, no previously existing MSFileState will be overwritten, and
with the status STATE_PRESENT, there won't be any request sent to the scanner.

Ref: DLP-5042